reviewd a couple of maker filter field descriptions

+++ userspace/libsinsp/utils.cpp
@@ -1014,7 +1014,7 @@ int32_t sinsp_numparser::parsed8(const string& str)
 
 uint32_t sinsp_numparser::parseu16(const string& str)
 {
-	uint32_t res;
+	uint16_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRIu16 "%c", &res, &temp) != 1)
@@ -1027,7 +1027,7 @@ uint32_t sinsp_numparser::parseu16(const string& str)
 
 int32_t sinsp_numparser::parsed16(const string& str)
 {
-	int32_t res;
+	int16_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRId16 "%c", &res, &temp) != 1)
